pyproject.toml
[tool.poetry]
name = "doki-backend"
version = "0.1.0"
description = "Doki backend (FastAPI) for sync engine"
authors = ["Doki Team <dev@doki>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "*"
uvicorn = "*"
httpx = "*"
pydantic = "*"
python-dotenv = "*"
google-auth = "*"
google-auth-oauthlib = "*"
google-api-python-client = "*"
supabase = "*"
itsdangerous = "^2.2.0"
cryptography = "^44.0.0"
google-cloud-secret-manager = "^2.25.0"

[tool.poetry.group.dev.dependencies]
# add dev tools here if needed (e.g., pytest, ruff)

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
